Added root check for chown
Added a check to ensure that the script is running as root before attempting to chown the well-known directory. It is not critical that the ownership be changed, and the script functions fine in common scenarios. This just removes the ugly error message:
chown: changing ownership of '<path>/.well-known/acme-challenge/<token>': Operation not permitted

(This may be undesirable in a selinux scenario where the user the script is running as has such ability, but that is an edge case). An alternative would be to redirect stderr on that one command, or check the exit code and print a message explaining that it's probably fine that it couldn't.